requirements.txt
allure_python_commons==2.9.45
jsonpath==0.82
PyMySQL==1.0.2
pytest==5.4.3
PyYAML==6.0
requests==2.26.0
